LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/cross-origin-cookies.https-expected.txt


FAIL speculation rules based prefetch should not use cookies for cross origin urls. assert_equals: expected (undefined) undefined but got (string) "1"
